+++ CyproApplication/src/main/java/com/cypro/repository/PrimaryKeyGeneratorDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.cypro.entity.PrimaryKeyGenerator;
 
-public interface PrimaryKeyGeneratorDAO extends JpaRepository<PrimaryKeyGenerator, Long>{
+public interface PrimaryKeyGeneratorDAO extends JpaRepository<PrimaryKeyGenerator, Integer>{
 
 }
